# Api_Swagger_Develop.yml
openapi: 3.0.1
info:
  title: KALM API
  description: API для системы оценки преподавателей с ролями, магазином, опросами и аналитикой.
  version: 1.0.1
servers:
  - url: http://localhost:5113/api
    description: Local development server
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Тип ошибки
          example: "validation_error"
        message:
          type: string
          description: Человекочитаемое сообщение
          example: "Неверные входные данные"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                description: Поле с ошибкой
              message:
                type: string
                description: Сообщение об ошибке
          nullable: true
        statusCode:
          type: integer
          description: HTTP-статус код
          example: 400
        timestamp:
          type: string
          format: date-time
          description: Время возникновения ошибки
          example: "2025-04-02T12:00:00Z"
      required: [error, message, statusCode, timestamp]
    UnauthorizedErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Тип ошибки
          example: "unauthorized"
        message:
          type: string
          description: Человекочитаемое сообщение
          example: "Неавторизованный доступ"
        statusCode:
          type: integer
          description: HTTP-статус код
          example: 401
      required: [error, message, statusCode]
    ServerErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Тип ошибки
          example: "server_error"
        message:
          type: string
          description: Человекочитаемое сообщение
          example: "Ошибка сервера"
        statusCode:
          type: integer
          description: HTTP-статус код
          example: 500
      required: [error, message, statusCode]
...
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 30
          pattern: '^[a-zA-Z0-9_]+$'
          example: "student123"
        password:
          type: string
          minLength: 6
          maxLength: 100
          example: "securePassword123"
      required: [username, password]
    LoginRequest:
      type: object
      properties:
        username:
          type: string
          example: "student123"
        password:
          type: string
          example: "securePassword123"
      required: [username, password]
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        user:
          $ref: "#/components/schemas/User"

    # ===== Users =====
    User:
      type: object
      properties:
        id:
          type: integer
          minimum: 0
          example: "10"
        username:
          type: string
          example: "student123"
        fullName:
          type: string
          nullable: true
          example: "Иванов И.И."
        role:
          type: string
          example: "Студент"
        balance:
          type: integer
          minimum: 0
          example: 100
        avatar:
          type: string
          nullable: true
          example: "https://example.com/avatar.jpg"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Регистрация нового пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        201:
          description: Успешная регистрация
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        400:
          description: |
            Неверные данные:
            - Пользователя не существует
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalidUsername:
                  value:
                    error: "user_do_not_exists"
                    message: "Пользователя не существует"
                    statusCode: 400
                    details:
                      - field: "username"
                        message: "Такого пользователя не существует"
        500:
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerErrorResponse"
  /auth/login:
    post:
      tags: [Auth]
      summary: Вход в систему
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        200:
          description: Успешный вход
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        401:
          description: Неверные учетные данные
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalidUserData:
                  value:
                    error: "invalid_user_data"
                    message: "Неверные данные пользователя"
                    statusCode: 401
              
        500:
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerErrorResponse"
  /auth/me:
    get:
      tags: [Auth]
      summary: Получить данные текущего пользователя
      security:
        - bearerAuth: []
      responses:
        200:
          description: Успешно
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        401:
          description: Неавторизованный доступ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedErrorResponse"
        500:
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerErrorResponse"

  # ==================== Users (CRUD) ====================
...